Updates to links to CDR and BioCreative datasets
It seems that the entire BioCreative website has moved 
from  
https://www.biocreative.org/
to 
https://biocreative.bioinformatics.udel.edu/

I have accordingly updated the links to CDR and BioCreative datasets which now point to the new domain.

Additionally, it seems that the new website itself has not updated the links to its resources (and still points to the non-functional biocreative.org website). Thus, it may be required to make this substitution for anyone wishing to browse the website.